# specs/001-build-an-ai/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: AI Schedule Counseling API
  version: 0.1.0
  description: >-
    Minimal REST contracts for conversational scheduling assistant.
    Implementation via Next.js API routes. All responses JSON unless otherwise noted.
servers:
  - url: https://example.com/api
    description: Placeholder base URL (Next.js deployment)
components:
  schemas:
    ErrorResponse:
      type: object
      required: [ok, code, message]
      properties:
        ok:
          type: boolean
          const: false
        code:
          type: string
          description: Stable machine error code
        message:
          type: string
...
    CalendarEvent:
      type: object
      required: [id, title, start, end, durationMinutes]
      properties:
        id: { type: string }
        title: { type: string }
        start: { type: string, format: date-time }
        end: { type: string, format: date-time }
        durationMinutes: { type: integer }
    ChangeItem:
      type: object
      required: [id, type, event, rationale]
      properties:
        id: { type: string }
        type: { type: string, enum: [add, move, remove, adjust] }
        event:
          type: object
          required: [title, start, end, durationMinutes]
          properties:
            title: { type: string }
            start: { type: string, format: date-time }
            end: { type: string, format: date-time }
            durationMinutes: { type: integer }
        targetEventId: { type: string, nullable: true }
        rationale: { type: string }
        accepted:
          { type: string, enum: [pending, accepted, rejected], nullable: true }
    Proposal:
      type: object
      required: [id, revision, changes, summary, sleepAssessment, status]
      properties:
        id: { type: string }
        revision: { type: integer, minimum: 1 }
        changes:
          type: array
          items: { $ref: '#/components/schemas/ChangeItem' }
          minItems: 1
        summary: { type: string }
        sleepAssessment:
          type: object
          required: [estimatedSleepHours, belowTarget]
          properties:
            estimatedSleepHours: { type: number }
            belowTarget: { type: boolean }
        status:
          { type: string, enum: [draft, pending, approved, applied, discarded] }
    GenerateProposalRequest:
      type: object
      required: [problemText, events, preferences]
      properties:
        problemText: { type: string }
        clarifications:
          type: array
          items: { type: string }
        events:
          type: array
          items: { $ref: '#/components/schemas/CalendarEvent' }
        preferences:
          type: object
          properties:
            sleepTargetHours: { type: number }
            priorities:
              type: array
              items: { type: string }
    GenerateProposalResponse:
      type: object
      required: [ok, proposal]
      properties:
        ok: { type: boolean, const: true }
        proposal: { $ref: '#/components/schemas/Proposal' }
...
paths:
...
  /proposal/generate:
    post:
      summary: Generate or update a schedule proposal
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GenerateProposalRequest' }
      responses:
        '200':
          description: Proposal generated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GenerateProposalResponse' }
        '422':
          description: Invalid model output
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...